+++ backend/src/test/java/com/pickpick/channel/ChannelSubscriptionServiceTest.java
@@ -1,4 +1,4 @@
-package com.pickpick.service;
+package com.pickpick.channel;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
